Time: 1 ms (100.00%), Space: 43.9 MB (29.64%) - LeetHub

+++ 0036-valid-sudoku/0036-valid-sudoku.java
@@ -1,11 +1,11 @@
 class Solution {
     static char cur;
     public boolean isValidSudoku(char[][] board) {
-        for (short i=0; i<9; i++) {
-            for (short j=0; j<9; j++) {
+        for (int i=0; i<9; i++) {
+            for (int j=0; j<9; j++) {
                 cur = board[i][j];
                 if (cur=='.') continue;
-                for (short t=1; t<9; t++) {
+                for (int t=1; t<9; t++) {
                     if (cur==board[(i+t)%9][j]) return false;
                     if (cur==board[i][(j+t)%9]) return false;
                 }
